+++ src/metrics/api_v1_urls.py
@@ -1,5 +1,5 @@
 from django.conf.urls import url
-import views
+import api_v1_views as views
 
 urlpatterns = [
     url(r'article/hw,ga/(?P<doi>[\.\w]+\/[\.\w]+)/$', views.api_article_metrics_mixed_source, name='api-article-metrics-mixed-source'),
